refactor(SlideControls)[OK-94] Add new components in SlideControls directory and new context
