+++ client/src/components/board/SlideControls/DeleteSlideModal.tsx
